move cctp code to cctp.rs
